Split up runner
